Prevent excessive retries
Reduce excessive retry attempts by setting the default retry count to 0 for HTTP 500-series errors, preventing retries in both the client application and the proxy side.

If you do not want to retry on the proxy side, please set the maximum retry count using the `max_retries` parameter when instantiating the proxy.